fix(Details): refactors Details styles
