+++ src/packages/solid/components/Metadata/Details.types.ts
@@ -15,69 +15,59 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import type { NodeProps, NodeStyles } from '@lightningjs/solid';
-import type { ComponentStyleConfig, NodeStyleSet, Tone } from 'types/types.js';
+import type { ComponentStyleConfig, NodeStyleSet, TextStyleSet, Tone } from 'types/types.js';
 import type { UIComponentProps } from 'types/interfaces.js';
+import type { BadgeProps } from '../Badge/Badge.types.js';
+import type { RatingProps } from './Rating.types.js';
 
-export interface CheckboxProps extends UIComponentProps, CheckboxStyleProperties {
+export interface DetailsProps extends UIComponentProps, DetailsStyleProperties {
   /**
-   * Indicates whether the checkbox is checked or unchecked.
-   * Setting this to `true` will check the checkbox, and setting it to `false` will uncheck it.
+   * an array of BadgeProps to render [Badges](?path=/docs/components-badge--docs)
    */
-  checked?: boolean;
+  badges?: BadgeProps[];
+  /**
+   * an array of RatingProps to render ratings
+   */
+  ratings?: RatingProps[];
+  /**
+   * text to display as details title
+   */
+  title?: string;
   children?: NodeProps['children'];
 }
 
-export interface CheckboxStyleProperties {
+export interface DetailsStyleProperties {
   /**
-   * color of background when checkbox is checked
+   * alignment of items in details
+   */
+  alignItems?: NodeStyles['alignItems'];
+  /**
+   * The space between items in the details
+   */
+  contentSpacing?: number;
+  /**
+   * The space between individual badges in the details
+   */
+  badgeContentSpacing?: number;
+  /**
+   * The space between individual ratings in the details
+   */
+  ratingContentSpacing?: number;
+  /**
+   * color of text in details
    *
    * core reference: {@link https://github.com/lightning-js/renderer/blob/aefee0064a48055b3cf8dd459396ad4996b68ef5/src/main-api/INode.ts#L121 INodeWritableProps.color}
    */
-  backgroundColorChecked?: NodeStyles['color']; // TODO clew uses strokeColor, but we currently don't account for nested properties (checked.color)
-  /**
-   * color of background when checkbox is unchecked
-   *
-   * core reference: {@link https://github.com/lightning-js/renderer/blob/aefee0064a48055b3cf8dd459396ad4996b68ef5/src/main-api/INode.ts#L121 INodeWritableProps.color}
-   */
-  backgroundColor?: NodeStyles['color'];
-  /**
-   * radius of the border of checkbox
-   */
-  radius?: NodeStyles['borderRadius'];
-  /**
-   * placement of content in checkbox
-   */
-  justifyContent?: NodeStyles['justifyContent'];
-  // /**
-  //  * color of border around checkbox
-  //  *
-  //  * core reference: {@link https://github.com/lightning-js/renderer/blob/aefee0064a48055b3cf8dd459396ad4996b68ef5/src/main-api/INode.ts#L121 INodeWritableProps.color}
-  //  */
-  // strokeColor?: NodeStyles['color']; // TODO clew uses strokeColor, but we currently don't account for nested properties (border.color)
-  /**
-   * width of border around checkbox
-   */
-  // strokeWidth?: NodeStyles['borderWidth']; // TODO clew uses strokeWidth, but we currently don't account for nested properties (border.width)
-  /**
-   * width of icon in the checkbox
-   *
-   * only applied to built-in icons, ignored if child component is passed
-   */
-  checkWidth?: NodeStyles['width'];
-  /**
-   * height of icon in the checkbox
-   *
-   * only applied to built-in icons, ignored if child component is passed
-   */
-  checkHeight?: NodeStyles['height'];
+  textColor?: NodeStyles['color'];
 }
 
-export interface CheckboxStyles {
+export interface DetailsStyles {
   tone: Tone;
-  Container: NodeStyleSet;
-  Icon: NodeStyleSet;
+  Container: NodeStyleSet<{
+    badgeContentSpacing: number;
+    contentSpacing: number;
+    ratingContentSpacing: number;
+  }>;
+  Text: TextStyleSet;
 }
-
-export type CheckboxModes = 'checked' | 'focus' | 'disabled';
-
-export type CheckboxConfig = ComponentStyleConfig<CheckboxStyleProperties, CheckboxModes>;
+export type DetailsConfig = ComponentStyleConfig<DetailsStyleProperties>;
